Worked on delete user features
- added form to delete the user
- added confirmation
- lost item form now uses user properly

+++ src/actions/User.ts
@@ -33,6 +33,10 @@ interface UserUpdateData {
     description?: string;
 }
 
+interface UserDeleteData {
+    confirmation?: string;
+}
+
 export async function createUser(user: NewUser) {
     try {
         await dbConnect();
@@ -94,7 +98,7 @@ export async function updateUser(userId: string, userData: UserUpdateData): Prom
     }
 }
 
-export async function deleteUser(userId: string): Promise<{ success?: string; error?: string }> {
+export async function deleteUser(userId: string, userData: UserDeleteData): Promise<{ success?: string; error?: string }> {
     // Check if the userId is a valid ObjectId before updating
     if (!Types.ObjectId.isValid(userId)) {
         return { error: "Invalid user ID." };
@@ -103,12 +107,20 @@ export async function deleteUser(userId: string): Promise<{ success?: string; er
     try {
         await dbConnect();
 
-        const deletedUser = await User.findByIdAndDelete(userId);
+        if (userData.confirmation !== "Confirm Deletion") {
+            return { error: "Confirmation text does not match." };
+        }
 
-        if (!deletedUser) {
+        const userToDelete = await User.findById(userId);
+
+        if (!userToDelete) {
             return { error: "User not found." };
         }
         
+        doLogout();
+        const deletedUser = await User.findByIdAndDelete(userId);
+
+
         return { success: "User deleted successfully." };
     } catch (e: any) {
         return { error: "Unable to delete user, please try again." };
